new feature: option in time machine to open on a specific sha  (#18)
feat: Add option to time machine function to open on a specific sha

The option is "commit_sha", if the sha is found in the file's history,
the time machine will be opened on the sha directly. If it is not found,
a message is shown and the time machine opens as it would normally.